use custom AttributeDict globally
